requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
docstring_parser==0.16
elevenlabs==1.55.0
fastapi==0.100.0
frozenlist==1.6.0
groq==0.20.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hypercorn==0.14.4
hyperframe==6.1.0
idna==3.10
instructor==1.3.3
jiter==0.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-cerebras==0.5.0
langchain-core==0.3.49
langchain-groq==0.3.1
langchain-openai==0.3.12
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
openai==1.69.0
orjson==3.10.16
packaging==24.2
priority==2.0.0
propcache==0.3.1
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.27.0
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
zstandard==0.23.0
